speech_api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
vosk==0.3.45
soundfile==0.12.1
librosa==0.10.1
numpy==1.24.3
requests==2.31.0
tqdm==4.66.1
pydantic==2.4.2
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
websockets==12.0
aiofiles==23.2.0